clean it all up
